Moved ConnectionErrorException to commons package
