+++ airbyte-db/db-lib/src/main/java/io/airbyte/db/mongodb/MongoDatabase.java
@@ -14,10 +14,10 @@ import com.mongodb.client.MongoClients;
 import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoCursor;
 import com.mongodb.client.MongoIterable;
+import io.airbyte.commons.exceptions.ConnectionErrorException;
 import io.airbyte.commons.functional.CheckedFunction;
 import io.airbyte.commons.util.MoreIterators;
 import io.airbyte.db.AbstractDatabase;
-import io.airbyte.db.exception.ConnectionErrorException;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
