+++ airbyte-integrations/bases/base-java/src/main/java/io/airbyte/integrations/base/ssh/SshWrappedSource.java
@@ -5,6 +5,7 @@
 package io.airbyte.integrations.base.ssh;
 
 import com.fasterxml.jackson.databind.JsonNode;
+import io.airbyte.commons.exceptions.ConnectionErrorException;
 import io.airbyte.commons.util.AutoCloseableIterator;
 import io.airbyte.commons.util.AutoCloseableIterators;
 import io.airbyte.integrations.base.AirbyteTraceMessageUtility;
@@ -42,7 +43,7 @@ public class SshWrappedSource implements Source {
   public AirbyteConnectionStatus check(final JsonNode config) throws Exception {
     try {
       return SshTunnel.sshWrap(config, hostKey, portKey, delegate::check);
-    } catch (final SshException e) {
+    } catch (final SshException | ConnectionErrorException e) {
       final String sshErrorMessage = "Could not connect with provided SSH configuration. Error: " + e.getMessage();
       AirbyteTraceMessageUtility.emitConfigErrorTrace(e, sshErrorMessage);
       return new AirbyteConnectionStatus()
